Add files via upload
This code is the core framework for the paper ‘Event-triggered two-part separation control of multiple autonomous underwater vehicles based on extended observer’.
This paper investigates the problem of regulating the stratigraphic isolation of multi-autonomous underwater vehicles (auvs) based on the ‘leader-follower’ framework. Considering the cooperative-competitive relationship between the following auvs and the effects of unknown external disturbances, a backtracking-based extended state observer is designed to mitigate these disturbances, and an event-triggered control scheme is designed to achieve two-part consensus control of the multi-auv system. A rigorous theoretical analysis shows that the system achieves asymptotic stability without zero behaviour under the proposed event-triggered control scheme. Finally, the effectiveness of this regulation strategy to achieve formation separation within a multi-auv is verified by numerical simulations, where the trajectory tracking errors of a single auv are clustered in a compact region close to the source, achieving structural convergence, and the absence of zero behaviour is also demonstrated.

Note: This is the beginning of a series of studies on which the authors' team will next consider introducing slip film control algorithms as well as other methods to improve the adaptability of multi-AUV systems to complex operating conditions. Only the framework of the core code can be provided, not the complete code. Interested colleagues can consider to modify the control method and design the appropriate parameters by themselves.